+++ src/libtiled/varianttomapconverter.cpp
@@ -33,6 +33,7 @@
 #include "tilelayer.h"
 #include "tileset.h"
 #include "tilesetmanager.h"
+#include "wangset.h"
 
 #include <QScopedPointer>
 
@@ -369,10 +370,86 @@ SharedTileset VariantToMapConverter::toTileset(const QVariant &variant)
         tile->setProperties(extractProperties(tileVar));
     }
 
+    // Read Wang sets
+    const QVariantList wangSetVariants = variantMap[QLatin1String("wangsets")].toList();
+    for (const QVariant &wangSetVariant : wangSetVariants) {
+        if (auto wangSet = toWangSet(wangSetVariant.toMap(), tileset.data()))
+            tileset->addWangSet(wangSet);
+        else
+            return SharedTileset();
+    }
+
     if (!mReadingExternalTileset)
         mGidMapper.insert(firstGid, tileset);
 
     return tileset;
+}
+
+WangSet *VariantToMapConverter::toWangSet(const QVariantMap &variantMap, Tileset *tileset)
+{
+    const QString name = variantMap[QLatin1String("name")].toString();
+    const int tile = variantMap[QLatin1String("tile")].toInt();
+
+    QScopedPointer<WangSet> wangSet { new WangSet(tileset, name, tile) };
+
+    wangSet->setProperties(extractProperties(variantMap));
+
+    const QVariantList edgeColorVariants = variantMap[QLatin1String("edgecolors")].toList();
+    for (const QVariant &edgeColorVariant : edgeColorVariants)
+        wangSet->addWangColor(toWangColor(edgeColorVariant.toMap(), true));
+
+    const QVariantList cornerColorVariants = variantMap[QLatin1String("cornercolors")].toList();
+    for (const QVariant &cornerColorVariant : cornerColorVariants)
+        wangSet->addWangColor(toWangColor(cornerColorVariant.toMap(), false));
+
+    const QVariantList wangTileVariants = variantMap[QLatin1String("wangtiles")].toList();
+    for (const QVariant &wangTileVariant : wangTileVariants) {
+        const QVariantMap wangTileVariantMap = wangTileVariant.toMap();
+
+        const int tileId = wangTileVariantMap[QLatin1String("tileid")].toInt();
+        const QVariantList wangIdVariant = wangTileVariantMap[QLatin1String("wangid")].toList();
+
+        WangId wangId;
+        bool ok = true;
+        for (int i = 0; i < 8 && ok; ++i)
+            wangId.setIndexColor(i, wangIdVariant[i].toUInt(&ok));
+
+        if (!ok || !wangSet->wangIdIsValid(wangId)) {
+            mError = QLatin1String("Invalid wangId given for tileId: ") + QString::number(tileId);
+            return nullptr;
+        }
+
+        const bool fH = wangTileVariantMap[QLatin1String("hflip")].toBool();
+        const bool fV = wangTileVariantMap[QLatin1String("vflip")].toBool();
+        const bool fA = wangTileVariantMap[QLatin1String("dflip")].toBool();
+
+        Tile *tile = tileset->findOrCreateTile(tileId);
+
+        WangTile wangTile(tile, wangId);
+        wangTile.setFlippedHorizontally(fH);
+        wangTile.setFlippedVertically(fV);
+        wangTile.setFlippedAntiDiagonally(fA);
+
+        wangSet->addWangTile(wangTile);
+    }
+
+    return wangSet.take();
+}
+
+QSharedPointer<WangColor> VariantToMapConverter::toWangColor(const QVariantMap &variantMap,
+                                                             bool isEdge)
+{
+    const QString name = variantMap[QLatin1String("name")].toString();
+    const QColor color = variantMap[QLatin1String("color")].toString();
+    const int imageId = variantMap[QLatin1String("tile")].toInt();
+    const qreal probability = variantMap[QLatin1String("probability")].toDouble();
+
+    return QSharedPointer<WangColor>::create(0,
+                                             isEdge,
+                                             name,
+                                             color,
+                                             imageId,
+                                             probability);
 }
 
 ObjectTemplate *VariantToMapConverter::toObjectTemplate(const QVariant &variant)
@@ -460,7 +537,7 @@ TileLayer *VariantToMapConverter::toTileLayer(const QVariantMap &variantMap)
     }
     mMap->setLayerDataFormat(layerDataFormat);
 
-    if (dataVariant.isValid()) {
+    if (dataVariant.isValid() && !dataVariant.isNull()) {
         if (!readTileLayerData(*tileLayer, dataVariant, layerDataFormat,
                                QRect(startX, startY, tileLayer->width(), tileLayer->height()))) {
             return nullptr;