+++ src/tiled/toolmanager.cpp
@@ -21,6 +21,7 @@
 #include "toolmanager.h"
 
 #include "abstracttool.h"
+#include "preferences.h"
 
 #include <QAction>
 #include <QShortcut>
@@ -40,8 +41,11 @@ ToolManager::ToolManager(QObject *parent)
     , mSelectEnabledToolPending(false)
 {
     mActionGroup->setExclusive(true);
-    connect(mActionGroup, SIGNAL(triggered(QAction*)),
-            this, SLOT(actionTriggered(QAction*)));
+    connect(mActionGroup, &QActionGroup::triggered,
+            this, &ToolManager::actionTriggered);
+
+    connect(Preferences::instance(), &Preferences::languageChanged,
+            this, &ToolManager::retranslateTools);
 }
 
 ToolManager::~ToolManager()
@@ -89,8 +93,8 @@ QAction *ToolManager::registerTool(AbstractTool *tool)
     } else {
         toolAction->setToolTip(tool->name());
     }
-
     toolAction->setEnabled(tool->isEnabled());
+
     mActionGroup->addAction(toolAction);
 
     connect(tool, SIGNAL(enabledChanged(bool)),
@@ -149,8 +153,13 @@ void ToolManager::retranslateTools()
         // Update the text, shortcut and tooltip of the action
         action->setText(tool->name());
         action->setShortcut(tool->shortcut());
-        action->setToolTip(QString(QLatin1String("%1 (%2)")).arg(
-                tool->name(), tool->shortcut().toString()));
+        if (!tool->shortcut().isEmpty()) {
+            action->setToolTip(
+                        QString(QLatin1String("%1 (%2)")).arg(tool->name(),
+                                                              tool->shortcut().toString()));
+        } else {
+            action->setToolTip(tool->name());
+        }
     }
 }
 
@@ -174,13 +183,14 @@ void ToolManager::createShortcuts(QWidget *parent)
             // because different tools may use the same shortcut.
             shortcut->setEnabled(action->isEnabled());
             connect(action, &QAction::changed, shortcut, [=]() {
+                shortcut->setKey(action->shortcut());
                 shortcut->setEnabled(action->isEnabled());
             });
 
             connect(shortcut, &QShortcut::activated, action, &QAction::trigger);
 
-            // Unset the shortcut from the action to avoid ambiguous overloads
-            action->setShortcut(QKeySequence());
+            // Limit the context of the shortcut to avoid ambiguous overloads
+            action->setShortcutContext(Qt::WidgetShortcut);
         }
     }
 }