+++ NEWS.md
@@ -1,4 +1,4 @@
-### Tiled 1.1.5 (.. April 2018)
+### Tiled 1.1.5 (25 April 2018)
 
 * Fixed erasing mode of the Terrain Brush
 * Fixed crash after editing a template
@@ -7,11 +7,14 @@
 * Fixed shortcuts for flipping and rotating objects (#1926)
 * Fixed dynamic retranslation of tools and tool actions
 * Fixed possible crash when undoing/redoing Wang color changes
+* Fixed handling of sub-properties in Object Type Editor (#1936)
+* Fixed crash when deleting an object right before dragging it (#1933)
 * Adjust Wang tile data when tileset column count changes (#1851)
 * Improved fill behavior in case of selection on infinite map (#1921)
 * Removed ability to hide tile collision objects (#1929)
 * Remove tile collision layer along with the last object (#1230)
 * JSON plugin: Made the reader more strict about object types (#1922)
+* JSON plugin: Added support for Wang sets
 
 ### Tiled 1.1.4 (28 March 2018)